backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pymupdf==1.24.14
openai==1.57.0
python-dotenv==1.0.1
pytesseract==0.3.13
pillow==11.0.0
aiofiles==24.1.0
